+++ docs/bookmarklets/opena11y-example.js
@@ -11892,11 +11892,12 @@
    * @param  {String}   defaultRole  - Default role of element if no role is defined
    * @param  {Object}   node         - dom element node
    * @param  {String}   ariaVersion  - Version of ARIA to use for roles, props and state info
+   *                                   (Values: "ARIA12" | "ARIA13")
    */
 
   class AriaInfo {
-    constructor (doc, hasRole, role, defaultRole, node, ariaVersion='1.2') {
-      if (ariaVersion === `1.3`) {
+    constructor (doc, hasRole, role, defaultRole, node, ariaVersion='ARIA12') {
+      if (ariaVersion === `ARIA13`) {
         propertyDataTypes = propertyDataTypes$1;
         designPatterns    = designPatterns$1;
       }
@@ -16056,7 +16057,7 @@
    */
 
   class DOMElement {
-    constructor (parentInfo, elementNode, ordinalPosition, ariaVersion='1.2') {
+    constructor (parentInfo, elementNode, ordinalPosition, ariaVersion='ARIA12') {
       const parentDomElement = parentInfo.domElement;
       const accNameDoc       = parentInfo.useParentDocForName ?
                                parentInfo.parentDocument :
@@ -17538,6 +17539,7 @@
   /* common.js */
 
   const common = {
+    aria13: ' (ARIA 1.3)',
     level: ['undefined', 'AAA', 'AA', 'undefined', 'A'],
     baseResult: ['undefined','P','H','MC','W','V'],
     baseResultLong: ['undefined','Pass','Hidden','Manual Check','Warning','Violation'],
@@ -28072,8 +28074,10 @@
       return cell;
     }
 
-    updateColumnCount (col) {
-      this.colCount = Math.max(this.colCount, col);
+    updateColumnCount (endColumn) {
+      if (!isNaN(endColumn) && endColumn > 0) {
+        this.colCount = Math.max(this.colCount, endColumn-1);
+      }
     }
 
     getRow(rowNumber, domElement=null) {
@@ -28510,7 +28514,7 @@
         case 'td':
           if (te) {
             tc = te.addCell(domElement);
-            te.updateColumnCount(tc);
+            te.updateColumnCount(tc.endColumn);
           }
           break;
 
@@ -28734,7 +28738,7 @@
    */
 
   class DOMCache {
-    constructor (startingDoc, startingElement, ariaVersion='1.2') {
+    constructor (startingDoc, startingElement, ariaVersion='ARIA12') {
       if (typeof startingElement !== 'object') {
         startingElement = startingDoc.body;
       }
@@ -38735,14 +38739,12 @@
    * @param  {String} title        - A title of the evaluation
    *                                 (typically the title of the document)
    * @param  {String} url          - The URL to the document
-   * @param  {String}  ariaVersion - Version of ARIA to use for roles,
-   *                                 props and state info
    *
    * @return see @desc
    */
 
   class EvaluationResult {
-    constructor (startingDoc, title, url, ariaVersion='1.2') {
+    constructor (startingDoc, title, url) {
 
       this.startingDoc = startingDoc;
       this.title       = title;
@@ -38750,7 +38752,7 @@
       this.ruleset     = '';
       this.level       = '';
       this.scopeFilter = '';
-      this.ariaVersion = ariaVersion;
+      this.ariaVersion = '1.2';
 
       this.date           = getFormattedDate();
       this.version        = VERSION;
@@ -38770,18 +38772,22 @@
      * @param  {String}  ruleset     - Set of rules to evaluate (values: A" | "AA" | "AAA")
      * @param  {String}  level       - WCAG Level (values: 'A', 'AA', 'AAA')
      * @param  {String}  scopeFilter - Filter rules by scope (values: "ALL" | "PAGE" | "WEBSITE")
+     * @param  {String}  ariaVersion - Version of ARIA used for validation rules
+     *                                 (values: 'ARIA12' | ARIA13")
      */
 
-    runWCAGRules (ruleset='WCAG21', level='AA', scopeFilter='ALL') {
+    runWCAGRules (ruleset='WCAG21', level='AA', scopeFilter='ALL', ariaVersion='AR!A12') {
 
       const startTime = new Date();
       debug$2.flag && debug$2.log(`[evaluateWCAG][    ruleset]: ${ruleset}`);
       debug$2.flag && debug$2.log(`[evaluateWCAG][      level]: ${level}`);
       debug$2.flag && debug$2.log(`[evaluateWCAG][scopeFilter]: ${scopeFilter}`);
+      debug$2.flag && debug$2.log(`[evaluateWCAG][ariaVersion]: ${ariaVersion}`);
 
       this.ruleset     = ruleset;
       this.level       = level;
       this.scopeFilter = scopeFilter;
+      this.ariaVersion = ariaVersion;
 
       const domCache      = new DOMCache(this.startingDoc, this.startingDoc.body, this.ariaVersion);
       this.allDomElements = domCache.allDomElements;
@@ -38814,13 +38820,14 @@
      * @param  {Array}   ruleList  - Array of rule IDs to include in the evaluation
      */
 
-    runRuleListRules (ruleList) {
+    runRuleListRules (ruleList, ariaVersion='AR!A12') {
       const startTime = new Date();
       debug$2.flag && debug$2.log(`[evaluateRuleList][ruleList]: ${ruleList}`);
 
       this.ruleset     = 'RULELIST';
+      this.ariaVersion = ariaVersion;
 
-      const domCache      = new DOMCache(this.startingDoc);
+      const domCache      = new DOMCache(this.startingDoc, ariaVersion);
       this.allDomElements = domCache.allDomElements;
       this.allRuleResults = [];
 
@@ -38844,12 +38851,13 @@
      * @desc Updates rule results array with results first step rules
      */
 
-    runFirstStepRules () {
+    runFirstStepRules (ariaVersion='AR!A12') {
       const startTime = new Date();
 
       this.ruleset     = 'FIRSTSTEP';
+      this.ariaVersion = ariaVersion;
 
-      const domCache      = new DOMCache(this.startingDoc);
+      const domCache      = new DOMCache(this.startingDoc, ariaVersion);
       this.allDomElements = domCache.allDomElements;
       this.allRuleResults = [];
 
@@ -39110,12 +39118,14 @@
      * @param  {String}  title       - Title of document being analyzed
      * @param  {String}  url         - URL of document being analyzed
      * @param  {Array}   ruleList    - Array of rule id to include in the evaluation
+     * @param  {String}  ariaVersion - Version of ARIA used for validation rules
+     *                                 Values: 'ARIA12' | 'ARIA13'
      */
 
-    evaluateRuleList (startingDoc, title='', url='',  ruleList = []) {
+    evaluateRuleList (startingDoc, title='', url='',  ruleList = [], ariaVersion='ARIA12') {
 
       const evaluationResult = new EvaluationResult(startingDoc, title, url);
-      evaluationResult.runRuleListRules(ruleList);
+      evaluationResult.runRuleListRules(ruleList, ariaVersion);
 
       // Debug features
       if (debug$1.flag) {
@@ -39145,12 +39155,14 @@
      * @param  {String}  ruleset     - Set of rules to evaluate (values: A" | "AA" | "AAA")
      * @param  {String}  level       - WCAG Level (values: 'A', 'AA', 'AAA')
      * @param  {String}  scopeFilter - Filter rules by scope (values: "ALL" | "PAGE" | "WEBSITE")
-     */
+     * @param  {String}  ariaVersion - Version of ARIA used for validation rules
+     *                                 Values: 'ARIA12' | 'ARIA13'
+    */
 
-    evaluateWCAG (startingDoc, title='', url='', ruleset='WCAG22', level='AAA', scopeFilter='ALL') {
+    evaluateWCAG (startingDoc, title='', url='', ruleset='WCAG22', level='AAA', scopeFilter='ALL', ariaVersion) {
 
       const evaluationResult = new EvaluationResult(startingDoc, title, url);
-      evaluationResult.runWCAGRules(ruleset, level, scopeFilter);
+      evaluationResult.runWCAGRules(ruleset, level, scopeFilter, ariaVersion);
 
       // Debug features
       if (debug$1.flag) {
@@ -39177,12 +39189,14 @@
      * @param  {Object}  startingDoc - Browser document object model (DOM) to be evaluated
      * @param  {String}  title       - Title of document being analyzed
      * @param  {String}  url         - url of document being analyzed
-     */
+     * @param  {String}  ariaVersion - Version of ARIA used for validation rules
+     *                                 Values: 'ARIA12' | 'ARIA13'
+    */
 
-    evaluateFirstStepRules (startingDoc, title='', url='') {
+    evaluateFirstStepRules (startingDoc, title='', url='', ariaVersion) {
 
       const evaluationResult = new EvaluationResult(startingDoc, title, url);
-      evaluationResult.runFirstStepRules();
+      evaluationResult.runFirstStepRules(ariaVersion);
 
       // Debug features
       if (debug$1.flag) {