+++ src/ainspector-content-script/content.js
@@ -55,6 +55,7 @@ function getEvaluationInfo(panelPort) {
     debug.log(`[getEvaluationInfo][        ruleset]: ${aiInfo.ruleset}`);
     debug.log(`[getEvaluationInfo][          level]: ${aiInfo.level}`);
     debug.log(`[getEvaluationInfo][    scopeFilter]: ${aiInfo.scopeFilter}`);
+    debug.log(`[getEvaluationInfo][    ariaVersion]: ${aiInfo.ariaVersion}`);
     debug.log(`[getEvaluationInfo][      highlight]: ${aiInfo.highlight}`);
     debug.log(`[getEvaluationInfo][       resultId]: ${aiInfo.resultId}`);
     debug.log(`[getEvaluationInfo][  highlightOnly]: ${aiInfo.highlightOnly}`);
@@ -70,18 +71,18 @@ function getEvaluationInfo(panelPort) {
   info.ruleset     = aiInfo.ruleset;
   info.level       = aiInfo.level;
   info.scopeFilter = aiInfo.scopeFilter;
-  info.evaluationLabel = getRulesetLabel(aiInfo.ruleset, aiInfo.level);
+  info.evaluationLabel = getRulesetLabel(aiInfo.ruleset, aiInfo.level, aiInfo.ariaVersion);
 
   switch(aiInfo.view) {
     case viewId.allRules:
       clearHighlights();
-      info.infoAllRules = getAllRulesInfo(aiInfo.ruleset, info.level, aiInfo.scopeFilter);
+      info.infoAllRules = getAllRulesInfo(aiInfo.ruleset, info.level, aiInfo.scopeFilter, aiInfo.ariaVersion);
       ruleResult = false;
       break;
 
     case viewId.ruleGroup:
       clearHighlights();
-      info.infoRuleGroup = getRuleGroupInfo(aiInfo.groupType, aiInfo.groupId, aiInfo.ruleset, info.level, aiInfo.scopeFilter);
+      info.infoRuleGroup = getRuleGroupInfo(aiInfo.groupType, aiInfo.groupId, aiInfo.ruleset, info.level, aiInfo.scopeFilter, aiInfo.ariaVersion);
       ruleResult = false;
       break;
 
@@ -94,7 +95,7 @@ function getEvaluationInfo(panelPort) {
           info.infoHighlight = true;
         }
       } else {
-        const evaluationResult  = evaluate(aiInfo.ruleset, aiInfo.level, aiInfo.scopeFilter);
+        const evaluationResult  = evaluate(aiInfo.ruleset, aiInfo.level, aiInfo.scopeFilter, aiInfo.ariaVersion);
         ruleResult = evaluationResult.getRuleResult(aiInfo.ruleId);
         info.infoRuleResult = getRuleResultInfo(ruleResult);
         highlightResults(ruleResult.getAllResultsArray(), aiInfo.highlight, aiInfo.resultId);